pyproject.toml
[tool.poetry]
name = "mylib"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
jupyter = "^1.0.0"
pandas = "^2.0.3"
seaborn = "^0.12.2"
tqdm = "^4.65.0"
sqlalchemy = "^2.0.23"
python-dotenv = "^1.0.0"
scikit-learn = "1.3.2"
hydra-core = "^1.3.2"
mlflow = "^2.9.2"
ydata-profiling = "^4.6.3"
lightgbm = "^4.2.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
black = {extras = ["jupyter"], version = "^23.7.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
